Basic input/output and subworkflow scattering tests.
There was a bunch of complexity related to simple input and output, default values, optional values, Any, union types, etc... to workflows and tools in Galaxy. This adds a bunch of conformance tests that isolates these basic input and output concerns.

There were also test cases for subworkflow scattering - this adds a couple of those.