Tests for specification of secondaryFiles in subdirectories of the job input document.
Other secondaryFile input object test specified files in the same directory as the job input document. To me it was not obvious if the locatin/path of the secondary inputs should have been relative to their parent input or relative to the job itself.

Unfortunately cwltool just picks up secondaryFiles as needed without them being specified, so I'm not absolutely certain the behavior is the intended behavior but it is my best guess.