Added CWLObjectType named union
This commit introduces a `CWLObjectType` named union to represent a
valid CWL object. Using a more precise union type instead of `Any`
allows for a smarter parsing from SALAD auto-generated parsers.
In particular, `CWLObjectType` is used to define the type of the
`default` directive in CWL `InputParameter` and `WorkflowStepInput`
fields. Plus, a `CWLInputFile` type is introduced to represent a CWL
jobfiles as `map<string, CWLObjectType>`.